refactor(controls-panel): extract reusable section component and move sections to layout
