--- src/features/coverage-planning/components/controls-panel/ObjectSection/ObjectSection.tsx
@@ -1,21 +1,0 @@
-import { useControlsPanelStore } from "@/stores/controlsPanelStore";
-import ControlsPanelSectionTitle from "@/components/ControlsPanelSectionTitle/ControlsPanelSectionTitle";
-import ObjectTypeSelect from "@/features/coverage-planning/components/controls-panel/ObjectSection/ObjectTypeSelect/ObjectTypeSelect";
-
-export default function ObjectSection() {
-  const isExpanded = useControlsPanelStore(
-    (state) => state.expandedSections["object"],
-  );
-
-  return (
-    <div className="flex flex-col">
-      <ControlsPanelSectionTitle sectionId="object" title="Object" />
-
-      {isExpanded && (
-        <div className="flex flex-col pl-1">
-          <ObjectTypeSelect />
-        </div>
-      )}
-    </div>
-  );
-}
