+++ src/layouts/workbench-window/controls-panel/algo-section/AlgoSection.tsx
@@ -1,0 +1,14 @@
+import ControlsPanelSection from "@/components/ControlsPanelSection/ControlsPanelSection";
+import AlgoSelect from "@/features/coverage-planning/components/controls-panel/AlgoSection/AlgoSelect/AlgoSelect";
+import PathWidthInput from "@/features/coverage-planning/components/controls-panel/AlgoSection/PathWidthInput/PathWidthInput";
+import PathOverlapInput from "@/features/coverage-planning/components/controls-panel/AlgoSection/PathOverlapInput/PathOverlapInput";
+
+export default function AlgoSection() {
+  return (
+    <ControlsPanelSection sectionId="algo" title="Coverage Path">
+      <AlgoSelect />
+      <PathWidthInput />
+      <PathOverlapInput />
+    </ControlsPanelSection>
+  );
+}
