+++ src/layouts/workbench-window/controls-panel/env-section/EnvSection.tsx
@@ -1,0 +1,12 @@
+import ControlsPanelSection from "@/components/ControlsPanelSection/ControlsPanelSection";
+import GlobalTypeSelection from "@/features/coverage-planning/components/controls-panel/EnvSection/GlobalTypeSelect/GlobalTypeSelect";
+import FormatSelection from "@/features/coverage-planning/components/controls-panel/EnvSection/FormatSelect/FormatSelect";
+
+export default function EnvSection() {
+  return (
+    <ControlsPanelSection sectionId="env" title="Environment">
+      <GlobalTypeSelection />
+      <FormatSelection />
+    </ControlsPanelSection>
+  );
+}
